refactor: rename useAppSettings to useSettings
