+++ src/components/theme-selector/ThemeSelector.test.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { STORAGE_KEY } from "@/hooks/useAppSettings";
+import { STORAGE_KEY } from "@/hooks/useSettings";
 import ThemeSelector from "./ThemeSelector";
 
 afterEach(() => {
